Move websocket manager to controllers
